Initial commit: a Daily calls dashboard SPA.
At present, the code has a server that talks to the rooms endpoints
of the Daily video API and a client that displays very basic info about
them in a table, with links to pages that would someday hold metric
metadata.

It's very busy with passing functions around which allow subcomponents
to modify the state of the 'App', i.e., which allow subcomponents to
trigger navigation to other 'pages'. It's also very busy with nested
ternary statements and the html`` function from preact, which has
made even this simple app into a bit of a mess.

Rather than try to clean this client code up, for now I'm going to try
a very different older-school approach and just route to pages stored
in templates on the server. I think it'll be a lot simpler for this
project.