Add call metrics temp storage + display.
This commit adds the nunjuck templating library (like jinja) and makes
the project into a few server-served template pages rather than a react
style SPA. If this project had more dynamic UI elements or more inter-
page communication and navigation, the Single Page approach might make
more sense, but for the deliverables, I feel this is far simpler.

The dashboard still lists the rooms and lets you create/delete/join
them, the metrics page now shows you all of the collected metrics for a
given room, and the call page has the actual daily call embedded and
captures+uploads the metrics for the viewing participant.

For now, the metrics are just stored in memory. When retrieved, they're
passed to a Vega-lite embed to render a layered chart with each session
dataset shown in different colours, and each type of data (send/recv
kbps and send/recv packet loss) indicated by different line-dashing and
mark-point-shapes. The raw data is also given in a table below.

Finally, errors (even async thrown ones!) should be safely caught and
sent out to the client, because of express-5's async error handling and
the custom error handler added.