Persist metrics in a sqlite3 db using knex.
I'm tired of all my metrics disappearing every time I save my files,
so let's pass a storage mechanism to the MetricsStorage class and
use it to persist the metrics to a db file. We'll also hook the db
calls up to the metrics rendering page, and delete a room's metrics
when we delete the room.

At this stage, we aren't handling creating the db and its schema
automatically -- it is assumed to exist in the form we need.