Small visual improvements to tables+metrics.
This commit styles the tables a little bit, formats the bps/kbps
metrics to a fixed 2-decimals, changes the link to the dashboard
to be a bit more descriptive, and enables tooltips on the vega chart.

The spec says,

> We are not worried about the UX or UI design at all. We know we're
> asking you to do a lot, and we'd prefer you spend any extra "polish"
> time you have on making your code as clear, modular, and
> appropriately commented as possible.

so... this really isn't pertinent, but I couldn't stand how the tables
looked and wanted to just take a few minutes to make things at least
a tiny bit easier on my eyes!