Shuffle everything around into their own homes.
This moves everything from server/index.js to src/<many folders and
files>. Only the actual app setup and running remains in the index, and
everything else is in their subfolders and files now. Some new tools
are also added in the process, like `app.db` as an app-wide database
context-holder, so that multiple different modules can use the db and
not need to require files or worry about re-making connections.

When I'm writing a small, greenfield project and don't know what bits
it'll actually include in the end, I often just agglomerate all of the
code in one place. When it's changing a lot, it's easiest if it's all
together and I can find everything just scrolling around, or searching
with a non-crossfile-search.

But once something is near a first complete version, and the pieces are
reasonably stable, they should have homes that somewhat semantically
describe their function. Not that this project structure is *perfect*,
far from it! It's also harder to read the full execution flow of the
app this way. But, it starts to build boundaries between components so
that they can't grow into each other over time by accident, and those
boundaries will help to force us (er, me I guess) into working toward
actually-good abstractions over time.

I normally might not make/keep earlier commits in a project this small
before reaching this stage. For the sake of showing my work, it's all
still there. And in a real or bigger project, I would likely never do a
big change-around like this -- it's disruptive and dangerous on a team
to move everything around! It also messes with git blame. In that kind
of situation, I'd usually make piecemeal changes over time in the
context of other work being done. I'd *also* be relying on tests...
which I haven't written *any* of yet for this project, and should have
done some of as I went along.