Keep app 'working' without Daily API Key.
Even if a key doesn't get passed, we should render what we can. If
you happen to know a /metrics/:room_name link and the server has
metrics for it, it should work properly even if you don't have the
key (or if the key were expired).

Similarly, the dashboard should render *something* nice in that case,
where it was previously just showing a json error message!