Add some integration tests.
These are all integration tests in that they are all testing the public
contracts of their components -- they mock things like the Daily API or
the DB's responses, and they test things like "what the toy metrics
collector API returns", or "what the database received after an API
call".

The view tests load the app and check for content in the html (in a
longer-lived project, I think parsing the html would be a better move,
but for now checking the text is fine), the api tests also load the app
and test the more "server"-y aspects of the project, and the metrics
storage tests just supply a mocked db and directly test their calls.

I struggled to find a good abstraction for mocking the db, so we've
ended up with mock-knex and directly checking the emitted sql and param
bindings, which I think would be fragile for any real length of time.
For now, it gives high confidence that the exact query I desire is
being emitted. I don't like the db tests, but I needed *something*.

No true unit tests were added. The DailyAPI class is mostly just a thin
wrapper around http calls and is not really worth testing -- the other
tests cover its connection points well.

One big block of untested code is the static js in the templates. It's
mostly 'too simple to test': some button events on the index page, the
vega chart on the metrics page, and the setInterval to call the metrics
upload function. Even the metrics upload function is just "call
getNetworkStats(), then upload the results", which would be a lot to
mock for little logic to test. There's still a fair bit of code here
overall, so it feels like this is a bit of a weak point in the design
chosen.